Rename integration component
