+++ web/src/components/molecules/Integration/IntegrationConnectModal/index.tsx
@@ -2,7 +2,7 @@ import styled from "@emotion/styled";
 
 import Button from "@reearth-cms/components/atoms/Button";
 import Modal from "@reearth-cms/components/atoms/Modal";
-import IntegrationCard from "@reearth-cms/components/molecules/Integration/IntegrationConnectModal/integrationCard";
+import IntegrationCard from "@reearth-cms/components/molecules/Integration/IntegrationConnectModal/IntegrationCard";
 import { useT } from "@reearth-cms/i18n";
 
 export interface Props {
